+++ DSL.Reqnroll.sln
@@ -1,4 +1,3 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34728.176
@@ -11,7 +10,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DSL.Examples", "DSL.Examples\DSL.Examples.csproj", "{014AE8D9-95BD-4FC9-9A27-ED59E9D0EBBB}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DSL.ReqnrollPlugin.Examples", "DSL.ReqnrollPlugin.Examples\DSL.ReqnrollPlugin.Examples.csproj", "{014AE8D9-95BD-4FC9-9A27-ED59E9D0EBBB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
